+++ src/Sylves.Test/Grid/Triangle/TrianglePrismCellCellTypeTest.cs
@@ -1,7 +1,11 @@
 ﻿using NUnit.Framework;
+using System;
+using System.Linq;
 #if UNITY
 using UnityEngine;
 #endif
+
+using static Sylves.Test.TestUtils;
 
 
 namespace Sylves.Test